+++ CHANGES.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixes a couple of bugs related to using non-default modules and criteria (#927)
+
 ## [0.12.1] - 2022-11-18
 
 ### Changed